requirements.txt
streamlit
streamlit-cropper
streamlit-image-coordinates
pdf2image
Pillow
numpy
pandas
openpyxl
ExtractTable
python-dotenv 